Add marked dates functionality
Added a new markedDates prop and a 'marked' selector, along with a
function that checks for each date string on the prop array, parses
it using the parseDateString() function and compares it to the day
state iterator. The 'selected' class selector has precedence over
the 'marked' selector.

Default color for 'marked' selector is #c2747d.